requirements.txt
paho-mqtt
python-systemd
systemd-python
